+++ src/backend/joanie/core/models/products.py
@@ -145,22 +145,18 @@ class Product(parler_models.TranslatableModel, BaseModel):
         an equivalent course run with null dates and hidden visibility.
         """
         site = Site.objects.get_current()
-        aggregate = self.target_course_runs.aggregate(
-            models.Min("start"),
-            models.Max("end"),
-            models.Max("enrollment_start"),
-            models.Min("enrollment_end"),
-        )
         resource_path = reverse("products-detail", kwargs={"id": self.id})
+        dates = self.get_equivalent_course_run_dates()
+
         return {
             "resource_link": f"https://{site.domain:s}{resource_path:s}",
             "catalog_visibility": visibility
-            or (enums.COURSE_AND_SEARCH if any(aggregate.values()) else enums.HIDDEN),
+            or (enums.COURSE_AND_SEARCH if any(dates.values()) else enums.HIDDEN),
             "languages": self.get_equivalent_course_run_languages(),
             # Get dates from aggregate
             **{
-                key.split("__")[0]: value.isoformat() if value else None
-                for key, value in aggregate.items()
+                key: value.isoformat() if value else None
+                for key, value in dates.items()
             },
         }
 
@@ -171,6 +167,24 @@ class Product(parler_models.TranslatableModel, BaseModel):
         ).distinct()
         # Go through a set for uniqueness of each language then return an ordered list
         return sorted(list(set(itertools.chain.from_iterable(languages))))
+
+    def get_equivalent_course_run_dates(self):
+        """
+        Return a dict of dates equivalent to course run dates
+        by aggregating dates of all target course runs as follows:
+        - start: Pick the earliest start date
+        - end: Pick the latest end date
+        - enrollment_start: Pick the latest enrollment start date
+        - enrollment_end: Pick the earliest enrollment end date
+        """
+        aggregate = self.target_course_runs.aggregate(
+            models.Min("start"),
+            models.Max("end"),
+            models.Max("enrollment_start"),
+            models.Min("enrollment_end"),
+        )
+
+        return {key.split("__")[0]: value for key, value in aggregate.items()}
 
     @staticmethod
     def get_equivalent_serialized_course_runs_for_products(
@@ -196,6 +210,14 @@ class Product(parler_models.TranslatableModel, BaseModel):
                 )
 
         return equivalent_course_runs
+
+    @property
+    def state(self):
+        """
+        Process the state of the product based on its equivalent course run dates.
+        """
+        dates = self.get_equivalent_course_run_dates()
+        return courses_models.CourseRun.compute_state(**dates)
 
     def clean(self):
         """